requirements.txt
PyPDF2 == 2.5.0
python-docx == 0.8.11
